fix type errors

+++ WebAnalytics.Core/Entities/DailyStats.cs
@@ -13,7 +13,7 @@ namespace WebAnalytics.Core.Entities
         public int TotalUsers { get; set; }
         public int TotalSessions { get; set; }
         public int TotalViews { get; set; }
-        public double AvgPerformance { get; set; }
+        public decimal AvgPerformance { get; set; }
         public DateTime LastUpdatedAt { get; set; } = DateTime.UtcNow;
     }
