+++ WebAnalytics.Core/Entities/RawData.cs
@@ -14,7 +14,7 @@ namespace WebAnalytics.Core.Entities
         public int Users { get; set; }
         public int Sessions { get; set; }
         public int Views { get; set; }
-        public double PerformanceScore { get; set; }
+        public decimal PerformanceScore { get; set; }
         public int LCPms { get; set; }
         public DateTime ReceivedAt { get; set; } = DateTime.UtcNow;
     }
